* commit '64825fa242c760f28da2d51a5e0ccd7d99df2d4d': (151 commits)
  s32g274ardb: Enable VR5510 PMIC for QSPI boot
  s32g274a: Skip VR5510 disablement if not in INIT_FS state
  s32g274a: Do not disable VR5510 watchdog on ATF bootflow
  s32g274a: Disable VR5510 watchdog
  s32g274ardb: Correct fdt_file
  fix sec-boot issues on u-boot 2020
  s32g274a: Add dts fix-up as part of LPDDR4 ERR050543
  sec-boot: fix cache issues
  sec-boot: automate u-boot secboot config
  sec-boot: add automatic sd layout secboot config
  sec-boot: remove dts specification in secboot config fragment
  sec-boot: add u-boot signing support
  s32r45-evb: Update ddr code for S32R45 platform
  s32g274a: Add ddr code with Errata for Temperature.
  campps32v2: rename defconfig to s32v234campp
  campps32v2: set fdt_file based on SoC id
  s32v: reworked lpddr2 config
  campps32v2: assign default mac and ip based on id
  pcie: s32v: do not force link to gen 1 in ep mode
  campps32v2: enable PCIE EP mode with external clock
  ...

+++ drivers/mmc/fsl_esdhc_imx.c
@@ -293,7 +293,7 @@ static int esdhc_setup_data(struct fsl_esdhc_priv *priv, struct mmc *mmc,
 		else
 			esdhc_write32(&regs->dsaddr, lower_32_bits(addr));
 #else
-		esdhc_write32(&regs->dsaddr, (u32)data->dest);
+		esdhc_write32(&regs->dsaddr, (u32)(uintptr_t)data->dest);
 #endif
 #endif
 	} else {
@@ -330,7 +330,7 @@ static int esdhc_setup_data(struct fsl_esdhc_priv *priv, struct mmc *mmc,
 		else
 			esdhc_write32(&regs->dsaddr, lower_32_bits(addr));
 #else
-		esdhc_write32(&regs->dsaddr, (u32)data->src);
+		esdhc_write32(&regs->dsaddr, (u32)(uintptr_t)data->src);
 #endif
 #endif
 	}
@@ -397,7 +397,7 @@ static void check_and_invalidate_dcache_range
 	else
 		start = lower_32_bits(addr);
 #else
-	start = (unsigned)data->dest;
+	start = (unsigned)(uintptr_t)data->dest;
 #endif
 	end = start + size;
 	invalidate_dcache_range(start, end);
